Add reason check to rating check

+++ src/main/java/com/ercanbeyen/bankingapplication/dto/RatingDto.java
@@ -17,7 +17,7 @@ public record RatingDto(
         LocalDateTime updatedAt,
         Integer year,
         @Range(min = 1, max = 5, message = "Rate should be between {min} and {max}")
-        Double rate,
+        Integer rate,
         RatingReason reason,
         String explanation) {
